+++ main.go
@@ -197,6 +197,12 @@ func main() {
 			os.Exit(1)
 		}
 	}
+
+	err = controller.SetupDashboardReconciler(mgr, conf)
+	if err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "Dashboard")
+		os.Exit(1)
+	}
 	//+kubebuilder:scaffold:builder
 
 	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {